Move document classes into workspacemanager
Align the document model with ADR-046 by relocating Document and
FileWatcherProcessor into the workspace manager package.

Update the affected imports and tests so the package move compiles
cleanly without changing behavior.
